+++ apps/frontend/src/libs/components/dashboard/components/current-plan/current-plan.tsx
@@ -1,7 +1,7 @@
 import { type FC } from "react";
 
-import { ExampleWithRemarksBig } from "~/assets/img/shared/illustrations/layouts/layouts.img.js";
 import { Link } from "~/libs/components/components.js";
+import { PlanStyle } from "~/libs/components/plan-styles/plan-style/plan-style.js";
 import { AppRoute } from "~/libs/enums/app-route.enum.js";
 
 import styles from "./styles.module.css";
@@ -11,11 +11,7 @@ const CurrentPlan: FC = () => {
 		<div className={styles["container"]}>
 			<h2 className={styles["title"]}>Current active plan</h2>
 			<div className={styles["plan-card"]}>
-				<img
-					alt="Current Plan Card"
-					className={styles["plan-image"]}
-					src={ExampleWithRemarksBig}
-				/>
+				<PlanStyle inputStyle="withremarks" />
 			</div>
 			<div className={styles["continue-button"]}>
 				<Link asButtonSize="small" asButtonVariant="primary" to={AppRoute.PLAN}>