+++ apps/frontend/src/libs/constants/weekdays.ts
@@ -1,0 +1,11 @@
+const WEEKDAYS = [
+	"Sunday",
+	"Monday",
+	"Tuesday",
+	"Wednesday",
+	"Thursday",
+	"Friday",
+	"Saturday",
+] as const;
+
+export { WEEKDAYS };