+++ apps/frontend/src/libs/types/plan/plan-view-option.type.ts
@@ -1,0 +1,3 @@
+type ViewOptions = "homepage" | "regular" | "selection";
+
+export { type ViewOptions };