+++ apps/frontend/src/libs/types/plan/task.type.ts
@@ -1,0 +1,10 @@
+type Task = {
+	description: string;
+	executionType?: string;
+	id: string;
+	isCompleted: boolean;
+	order?: number;
+	title?: string;
+};
+
+export { type Task };