+++ apps/frontend/src/libs/types/types.ts
@@ -3,7 +3,12 @@ export { type ButtonSize } from "./button-size.type.js";
 export { type ButtonVariant } from "./button-variant.type.js";
 export { type DownloadOptions, type Html2CanvasFunction } from "./pdf.type.js";
 export { type PlanEditForm } from "./plan-edit-form.type.js";
-export { type DayTopic, type PlanActivity, type PlanDay } from "./plan.type.js";
+export { type DayTopic, type PlanActivity, type Plan-Day } from "./plan.type.js";
+export { type PlanDay } from "./plan/plan-day.type.js";
+export { type PlanStyleOption } from "./plan/plan-style-option.type.js";
+export { type ViewOptions } from "./plan/plan-view-option.type.js";
+export { type Plan } from "./plan/plan.type.js";
+export { type Task } from "./plan/task.type.js";
 export {
 	type CheckboxQuestionProperties,
 	type MixedAnswer,