+++ apps/frontend/src/pages/home/components/visual-layouts/visual-layouts.tsx
@@ -3,10 +3,14 @@ import {
 	StarsYellow03,
 } from "~/assets/img/shared/shapes/shapes.img.js";
 import { DecorativeImage } from "~/libs/components/components.js";
+import { PlanStyle } from "~/libs/components/plan-styles/plan-style/plan-style.js";
 import { getClassNames } from "~/libs/helpers/get-class-names.js";
+import { type ViewOptions } from "~/libs/types/types.js";
 
 import { layoutExamples } from "./libs/constants.js";
 import styles from "./styles.module.css";
+
+const PLAN_VIEW_OPTION: ViewOptions = "homepage";
 
 const VisualLayouts: React.FC = () => {
 	return (
@@ -43,7 +47,10 @@ const VisualLayouts: React.FC = () => {
 								className={styles["layout-card"]}
 								key={example.id}
 							>
-								<DecorativeImage src={example.img} />
+								<PlanStyle
+									inputStyle={example.planStyle}
+									view={PLAN_VIEW_OPTION}
+								/>
 								<h5>{example.title}</h5>
 							</li>
 						))}